+++ src/frontend/packages/cloud-foundry/src/cloud-foundry-package.module.ts
@@ -5,10 +5,7 @@ import { MDAppModule } from '../../core/src/core/md.module';
 import { SharedModule } from '../../core/src/shared/shared.module';
 import { EntityCatalogModule } from '../../store/src/entity-catalog.module';
 import { generateCFEntities } from './cf-entity-generator';
-import { ApplicationsModule } from './features/applications/applications.module';
-import { CloudFoundryModule } from './features/cloud-foundry/cloud-foundry.module';
-import { ServiceCatalogModule } from './features/service-catalog/service-catalog.module';
-import { ServicesModule } from './features/services/services.module';
+import { CloudFoundryRoutingModule } from './cloud-foundry-routing.module';
 import { CloudFoundrySharedModule } from './shared/cf-shared.module';
 import { CfUserService } from './shared/data-services/cf-user.service';
 import { CloudFoundryService } from './shared/data-services/cloud-foundry.service';
@@ -24,13 +21,10 @@ import { cfCurrentUserPermissionsService } from './user-permissions/cf-user-perm
     CommonModule,
     SharedModule,
     MDAppModule,
-    ApplicationsModule,
-    CloudFoundryModule,
-    ServiceCatalogModule,
-    ServicesModule,
     CloudFoundryStoreModule,
     // FIXME: Ensure that anything lazy loaded is not included here - #3675
     CloudFoundrySharedModule,
+    CloudFoundryRoutingModule
   ],
   providers: [
     ...cfCurrentUserPermissionsService,
