+++ src/frontend/packages/cloud-foundry/src/features/cloud-foundry/cloud-foundry-section.routing.ts
@@ -427,4 +427,4 @@ const cloudFoundry: Routes = [{
 @NgModule({
   imports: [RouterModule.forChild(cloudFoundry)]
 })
-export class CloudFoundryRoutingModule { }
+export class CloudFoundrySectionRoutingModule { }
