+++ src/frontend/packages/cloud-foundry/src/features/cloud-foundry/users/manage-users/cf-roles.service.spec.ts
@@ -1,12 +1,12 @@
+import { HttpClientModule } from '@angular/common/http';
 import { inject, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 
 import { SharedModule } from '../../../../../../core/src/shared/shared.module';
 import { generateCfStoreModules } from '../../../../../test-framework/cloud-foundry-endpoint-service.helper';
 import { CfUserService } from '../../../../shared/data-services/cf-user.service';
-import { CloudFoundryModule } from '../../cloud-foundry.module';
+import { CloudFoundrySectionModule } from '../../cloud-foundry-section.module';
 import { CfRolesService } from './cf-roles.service';
-import { HttpClientModule } from '@angular/common/http';
 
 describe('CfRolesService', () => {
   beforeEach(() => {
@@ -14,7 +14,7 @@ describe('CfRolesService', () => {
       imports: [
         ...generateCfStoreModules(),
         SharedModule,
-        CloudFoundryModule,
+        CloudFoundrySectionModule,
         HttpClientModule,
         RouterTestingModule
       ],
