+++ src/frontend/packages/cloud-foundry/src/features/service-catalog/service-catalog.module.ts
@@ -1,4 +1,4 @@
-import { CommonModule } from '@angular/common';
+import { CommonModule, DatePipe } from '@angular/common';
 import { NgModule } from '@angular/core';
 
 import { CoreModule } from '../../../../core/src/core/core.module';
@@ -33,6 +33,9 @@ import { ServiceTabsBaseComponent } from './service-tabs-base/service-tabs-base.
   ],
   exports: [
     ServiceTabsBaseComponent,
+  ],
+  providers: [
+    DatePipe
   ]
 })
 export class ServiceCatalogModule { }
