+++ src/frontend/packages/cloud-foundry/src/shared/cf-shared.module.ts
@@ -208,9 +208,6 @@ import { CfUserPermissionDirective } from './directives/cf-user-permission/cf-us
 import { ApplicationStateService } from './services/application-state.service';
 import { CloudFoundryUserProvidedServicesService } from './services/cloud-foundry-user-provided-services.service';
 
-// tslint:disable:max-line-length
-// tslint:enable:max-line-length
-
 const cfListTableCells: Type<TableCellCustom<any>>[] = [
   TableCellServiceInstanceAppsAttachedComponent,
   TableCellServiceComponent,
@@ -266,8 +263,6 @@ const cfListCards: Type<CardCell<any>>[] = [
   ServiceInstanceCardComponent,
   UserProvidedServiceInstanceCardComponent,
 ];
-
-// listTableCells.push();
 
 @NgModule({
   imports: [
